log message checing and expose api to extend it
